+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Various documentation of the CRD ([#510]).
 - Helm: support labels in values.yaml ([#528]).
+- Delta Lake connector ([#531]).
 
 ### Fixed
 
@@ -16,6 +17,7 @@ All notable changes to this project will be documented in this file.
 [#510]: https://github.com/stackabletech/trino-operator/pull/510
 [#526]: https://github.com/stackabletech/trino-operator/pull/526
 [#528]: https://github.com/stackabletech/trino-operator/pull/528
+[#531]: https://github.com/stackabletech/trino-operator/pull/531
 
 ## [23.11.0] - 2023-11-24